feat: v2 de CRUD Produto

+++ src/main/java/com/gerencia_restaurante/application/port/in/CadastrarProduto.java
@@ -17,5 +17,5 @@ public record CadastrarProduto (
 
     @NotNull(message = "Preço mínimo obrigatório")
     @DecimalMin(value = "0.01", message = "Preço não pode ser zerado")
-    BigDecimal precoProduto
+    Double precoProduto
     ){}
